calculate_block_nesting {skpr}	R Documentation

Determine Nesting Level of Blocks

Description

Calculates if a block is fully nested within another block, and what the highest level of nesting is. 1 indicates the block isn't nested within another block–just within the intercept.

Usage

calculate_block_nesting(blockgroups, blockstructure)

Value

Vector of numbers indicating the hierarchy

[Package skpr version 1.9.2 Index]