calculatepower {skpr}	R Documentation

Calculate Power

Description

Calculates the power of the model given the non-centrality parameter

Usage

calculatepower(X, L, lambda, alpha, degrees_of_freedom)

Arguments

X
The model matrix

L
The parameter matrix/vector

lambda
The non-centrality parameter for the F test

alpha
the specified type-I error

degrees_of_freedom
The number of degrees of freedom.

Value

The power for a given parameter L, given the

[Package skpr version 1.9.2 Index]