genparammatrix {skpr}	R Documentation

Generate Parameter Matrix

Description

Generates the parameter matrix L to isolate the levels of interest in the calculation of power

Usage

genparammatrix(parameters, levels, g)

Arguments

parameters
Number of parameters total in model.

levels
Number of levels in parameter of interest

g
Number of levels/parameters preceeding parameter of interest

Value

The parameter vector Q isolating the levels of parameter of interest

[Package skpr version 1.9.2 Index]