normalize_design {skpr}	R Documentation

Normalize Design

Description

Normalizes the numeric columns in the design to -1 to 1. This is important to do if your model has interaction or polynomial terms, as these terms can introduce multi-collinearity and standardizing the numeric columns can reduce this problem.

Normalizes the numeric columns in the design to -1 to 1. This is important to do if your model has interaction or polynomial terms, as these terms can introduce multi-collinearity and standardizing the numeric columns can reduce this problem.

Usage

normalize_design(design, augmented = NULL)

normalize_design(design, augmented = NULL)

Arguments

design
The design matrix.

augmented
Default NULL. If

Value

Normalized design matrix

Normalized run matrix

Examples

#Normalize a design
if(skpr:::run_documentation()) {
cand_set = expand.grid(temp = c(100,300,500),
                      altitude = c(10000,20000),
                      offset = seq(-10,-5,by=1),
                      type = c("A","B", "C"))
design = gen_design(cand_set, ~., 24)

#Un-normalized design
design
}
if(skpr:::run_documentation()) {
#Normalized design
normalize_design(design)
}

[Package skpr version 1.9.2 Index]