Print evaluation information

Prints design evaluation information below the data.frame of power values

Note: If options("skpr.ANSI") is `NULL` or `TRUE`, ANSI codes will be used during printing to prettify the output. If this is `FALSE`, only ASCII will be used.

Usage

# S3 method for skpr_power_curve_output
print(x, ...)

Arguments

x

The x of the evaluation functions in skpr

...

Additional arguments.

Examples

#Generate/evaluate a design and print its information
factorialcoffee = expand.grid(cost = c(1, 2),
                              type = as.factor(c("Kona",
                                                 "Colombian",
                                                 "Ethiopian",
                                                 "Sumatra")),
                              size = as.factor(c("Short",
                                                 "Grande",
                                                 "Venti")))

coffee_curves = calculate_power_curves(candidateset = factorialcoffee,
                                      model = ~(cost + size + type)^2,
                                      trials = 30:40, plot_results = FALSE)
coffee_curves
#>       parameter            type     power trials effectsize random_seed
#> 1   (Intercept)    effect.power 0.6826885     30          1         123
#> 2          cost    effect.power 0.6799517     30          1         123
#> 3          size    effect.power 0.3791064     30          1         123
#> 4          type    effect.power 0.4506179     30          1         123
#> 5     cost:size    effect.power 0.3790878     30          1         123
#> 6     cost:type    effect.power 0.4478528     30          1         123
#> 7     size:type    effect.power 0.6245632     30          1         123
#> 8   (Intercept) parameter.power 0.6826885     30          1         123
#> 9          cost parameter.power 0.6799517     30          1         123
#> 10        size1 parameter.power 0.4052233     30          1         123
#> 11        size2 parameter.power 0.4027148     30          1         123
#> 12        type1 parameter.power 0.2798727     30          1         123
#> 13        type2 parameter.power 0.2821415     30          1         123
#> 14        type3 parameter.power 0.2983430     30          1         123
#> 15   cost:size1 parameter.power 0.4020228     30          1         123
#> 16   cost:size2 parameter.power 0.4042969     30          1         123
#> 17   cost:type1 parameter.power 0.2809834     30          1         123
#> 18   cost:type2 parameter.power 0.2809834     30          1         123
#> 19   cost:type3 parameter.power 0.2961708     30          1         123
#> 20  size1:type1 parameter.power 0.1736032     30          1         123
#> 21  size2:type1 parameter.power 0.1602483     30          1         123
#> 22  size1:type2 parameter.power 0.1605181     30          1         123
#> 23  size2:type2 parameter.power 0.1607885     30          1         123
#> 24  size1:type3 parameter.power 0.1641624     30          1         123
#> 25  size2:type3 parameter.power 0.1819501     30          1         123
#> 26  (Intercept)    effect.power 0.6997660     31          1         123
#> 27         cost    effect.power 0.6957588     31          1         123
#> 28         size    effect.power 0.3911743     31          1         123
#> 29         type    effect.power 0.4713329     31          1         123
#> 30    cost:size    effect.power 0.3857370     31          1         123
#> 31    cost:type    effect.power 0.4638559     31          1         123
#> 32    size:type    effect.power 0.6635006     31          1         123
#> 33  (Intercept) parameter.power 0.6997660     31          1         123
#> 34         cost parameter.power 0.6957588     31          1         123
#> 35        size1 parameter.power 0.4167653     31          1         123
#> 36        size2 parameter.power 0.4182993     31          1         123
#> 37        type1 parameter.power 0.2862835     31          1         123
#> 38        type2 parameter.power 0.3073151     31          1         123
#> 39        type3 parameter.power 0.3034260     31          1         123
#> 40   cost:size1 parameter.power 0.4115382     31          1         123
#> 41   cost:size2 parameter.power 0.4115382     31          1         123
#> 42   cost:type1 parameter.power 0.2864825     31          1         123
#> 43   cost:type2 parameter.power 0.3017406     31          1         123
#> 44   cost:type3 parameter.power 0.3017406     31          1         123
#> 45  size1:type1 parameter.power 0.1625825     31          1         123
#> 46  size2:type1 parameter.power 0.1621114     31          1         123
#> 47  size1:type2 parameter.power 0.1785585     31          1         123
#> 48  size2:type2 parameter.power 0.1674038     31          1         123
#> 49  size1:type3 parameter.power 0.1664224     31          1         123
#> 50  size2:type3 parameter.power 0.1850129     31          1         123
#> 51  (Intercept)    effect.power 0.7155349     32          1         123
#> 52         cost    effect.power 0.7114742     32          1         123
#> 53         size    effect.power 0.4042147     32          1         123
#> 54         type    effect.power 0.5010903     32          1         123
#> 55    cost:size    effect.power 0.4026865     32          1         123
#> 56    cost:type    effect.power 0.4843502     32          1         123
#> 57    size:type    effect.power 0.6122168     32          1         123
#> 58  (Intercept) parameter.power 0.7155349     32          1         123
#> 59         cost parameter.power 0.7114742     32          1         123
#> 60        size1 parameter.power 0.4200557     32          1         123
#> 61        size2 parameter.power 0.4325214     32          1         123
#> 62        type1 parameter.power 0.3205577     32          1         123
#> 63        type2 parameter.power 0.3205577     32          1         123
#> 64        type3 parameter.power 0.3153125     32          1         123
#> 65   cost:size1 parameter.power 0.4184315     32          1         123
#> 66   cost:size2 parameter.power 0.4313056     32          1         123
#> 67   cost:type1 parameter.power 0.3069861     32          1         123
#> 68   cost:type2 parameter.power 0.3069861     32          1         123
#> 69   cost:type3 parameter.power 0.3059611     32          1         123
#> 70  size1:type1 parameter.power 0.1700717     32          1         123
#> 71  size2:type1 parameter.power 0.1845655     32          1         123
#> 72  size1:type2 parameter.power 0.1700717     32          1         123
#> 73  size2:type2 parameter.power 0.1845655     32          1         123
#> 74  size1:type3 parameter.power 0.1815880     32          1         123
#> 75  size2:type3 parameter.power 0.1836763     32          1         123
#> 76  (Intercept)    effect.power 0.7387166     33          1         123
#> 77         cost    effect.power 0.7252330     33          1         123
#> 78         size    effect.power 0.4248411     33          1         123
#> 79         type    effect.power 0.5182235     33          1         123
#> 80    cost:size    effect.power 0.4213750     33          1         123
#> 81    cost:type    effect.power 0.5011963     33          1         123
#> 82    size:type    effect.power 0.6893940     33          1         123
#> 83  (Intercept) parameter.power 0.7387166     33          1         123
#> 84         cost parameter.power 0.7252330     33          1         123
#> 85        size1 parameter.power 0.4426127     33          1         123
#> 86        size2 parameter.power 0.4459904     33          1         123
#> 87        type1 parameter.power 0.3234550     33          1         123
#> 88        type2 parameter.power 0.3157942     33          1         123
#> 89        type3 parameter.power 0.3141142     33          1         123
#> 90   cost:size1 parameter.power 0.4392456     33          1         123
#> 91   cost:size2 parameter.power 0.4392456     33          1         123
#> 92   cost:type1 parameter.power 0.3106586     33          1         123
#> 93   cost:type2 parameter.power 0.3106586     33          1         123
#> 94   cost:type3 parameter.power 0.3106586     33          1         123
#> 95  size1:type1 parameter.power 0.1849669     33          1         123
#> 96  size2:type1 parameter.power 0.1877676     33          1         123
#> 97  size1:type2 parameter.power 0.1705872     33          1         123
#> 98  size2:type2 parameter.power 0.1920915     33          1         123
#> 99  size1:type3 parameter.power 0.1848221     33          1         123
#> 100 size2:type3 parameter.power 0.1711518     33          1         123
#> 101 (Intercept)    effect.power 0.7562022     34          1         123
#> 102        cost    effect.power 0.7376450     34          1         123
#> 103        size    effect.power 0.4313392     34          1         123
#> 104        type    effect.power 0.5370600     34          1         123
#> 105   cost:size    effect.power 0.4260546     34          1         123
#> 106   cost:type    effect.power 0.5185534     34          1         123
#> 107   size:type    effect.power 0.6996227     34          1         123
#> 108 (Intercept) parameter.power 0.7562022     34          1         123
#> 109        cost parameter.power 0.7376450     34          1         123
#> 110       size1 parameter.power 0.4515092     34          1         123
#> 111       size2 parameter.power 0.4549005     34          1         123
#> 112       type1 parameter.power 0.3195412     34          1         123
#> 113       type2 parameter.power 0.3168617     34          1         123
#> 114       type3 parameter.power 0.3453909     34          1         123
#> 115  cost:size1 parameter.power 0.4464771     34          1         123
#> 116  cost:size2 parameter.power 0.4464771     34          1         123
#> 117  cost:type1 parameter.power 0.3149170     34          1         123
#> 118  cost:type2 parameter.power 0.3149170     34          1         123
#> 119  cost:type3 parameter.power 0.3315543     34          1         123
#> 120 size1:type1 parameter.power 0.1946197     34          1         123
#> 121 size2:type1 parameter.power 0.1734524     34          1         123
#> 122 size1:type2 parameter.power 0.1736798     34          1         123
#> 123 size2:type2 parameter.power 0.1922518     34          1         123
#> 124 size1:type3 parameter.power 0.1983996     34          1         123
#> 125 size2:type3 parameter.power 0.1978415     34          1         123
#> 126 (Intercept)    effect.power 0.7759456     35          1         123
#> 127        cost    effect.power 0.7500965     35          1         123
#> 128        size    effect.power 0.4634431     35          1         123
#> 129        type    effect.power 0.5550627     35          1         123
#> 130   cost:size    effect.power 0.4438808     35          1         123
#> 131   cost:type    effect.power 0.5352030     35          1         123
#> 132   size:type    effect.power 0.7196053     35          1         123
#> 133 (Intercept) parameter.power 0.7759456     35          1         123
#> 134        cost parameter.power 0.7500965     35          1         123
#> 135       size1 parameter.power 0.4683468     35          1         123
#> 136       size2 parameter.power 0.4863533     35          1         123
#> 137       type1 parameter.power 0.3279723     35          1         123
#> 138       type2 parameter.power 0.3498501     35          1         123
#> 139       type3 parameter.power 0.3415317     35          1         123
#> 140  cost:size1 parameter.power 0.4536851     35          1         123
#> 141  cost:size2 parameter.power 0.4664493     35          1         123
#> 142  cost:type1 parameter.power 0.3190901     35          1         123
#> 143  cost:type2 parameter.power 0.3356799     35          1         123
#> 144  cost:type3 parameter.power 0.3356799     35          1         123
#> 145 size1:type1 parameter.power 0.1755796     35          1         123
#> 146 size2:type1 parameter.power 0.1894879     35          1         123
#> 147 size1:type2 parameter.power 0.1999700     35          1         123
#> 148 size2:type2 parameter.power 0.2031490     35          1         123
#> 149 size1:type3 parameter.power 0.2025563     35          1         123
#> 150 size2:type3 parameter.power 0.2070205     35          1         123
#> 151 (Intercept)    effect.power 0.8017265     36          1         123
#> 152        cost    effect.power 0.7627026     36          1         123
#> 153        size    effect.power 0.4805044     36          1         123
#> 154        type    effect.power 0.5779224     36          1         123
#> 155   cost:size    effect.power 0.4613127     36          1         123
#> 156   cost:type    effect.power 0.5512069     36          1         123
#> 157   size:type    effect.power 0.7281702     36          1         123
#> 158 (Intercept) parameter.power 0.8017265     36          1         123
#> 159        cost parameter.power 0.7627026     36          1         123
#> 160       size1 parameter.power 0.4990130     36          1         123
#> 161       size2 parameter.power 0.4876976     36          1         123
#> 162       type1 parameter.power 0.3614727     36          1         123
#> 163       type2 parameter.power 0.3614727     36          1         123
#> 164       type3 parameter.power 0.3641507     36          1         123
#> 165  cost:size1 parameter.power 0.4733657     36          1         123
#> 166  cost:size2 parameter.power 0.4733657     36          1         123
#> 167  cost:type1 parameter.power 0.3397409     36          1         123
#> 168  cost:type2 parameter.power 0.3397409     36          1         123
#> 169  cost:type3 parameter.power 0.3397409     36          1         123
#> 170 size1:type1 parameter.power 0.2038731     36          1         123
#> 171 size2:type1 parameter.power 0.2033581     36          1         123
#> 172 size1:type2 parameter.power 0.1979074     36          1         123
#> 173 size2:type2 parameter.power 0.2033581     36          1         123
#> 174 size1:type3 parameter.power 0.2025072     36          1         123
#> 175 size2:type3 parameter.power 0.1980670     36          1         123
#> 176 (Intercept)    effect.power 0.7864642     37          1         123
#> 177        cost    effect.power 0.7790389     37          1         123
#> 178        size    effect.power 0.4924130     37          1         123
#> 179        type    effect.power 0.5814906     37          1         123
#> 180   cost:size    effect.power 0.4648473     37          1         123
#> 181   cost:type    effect.power 0.5733489     37          1         123
#> 182   size:type    effect.power 0.7515928     37          1         123
#> 183 (Intercept) parameter.power 0.7864642     37          1         123
#> 184        cost parameter.power 0.7790389     37          1         123
#> 185       size1 parameter.power 0.5102800     37          1         123
#> 186       size2 parameter.power 0.5065711     37          1         123
#> 187       type1 parameter.power 0.3462992     37          1         123
#> 188       type2 parameter.power 0.3655256     37          1         123
#> 189       type3 parameter.power 0.3462992     37          1         123
#> 190  cost:size1 parameter.power 0.4819740     37          1         123
#> 191  cost:size2 parameter.power 0.4819740     37          1         123
#> 192  cost:type1 parameter.power 0.3445838     37          1         123
#> 193  cost:type2 parameter.power 0.3445838     37          1         123
#> 194  cost:type3 parameter.power 0.3445838     37          1         123
#> 195 size1:type1 parameter.power 0.2095725     37          1         123
#> 196 size2:type1 parameter.power 0.2025349     37          1         123
#> 197 size1:type2 parameter.power 0.2087190     37          1         123
#> 198 size2:type2 parameter.power 0.2006279     37          1         123
#> 199 size1:type3 parameter.power 0.2020649     37          1         123
#> 200 size2:type3 parameter.power 0.2080426     37          1         123
#> 201 (Intercept)    effect.power 0.8131992     38          1         123
#> 202        cost    effect.power 0.7962184     38          1         123
#> 203        size    effect.power 0.5168405     38          1         123
#> 204        type    effect.power 0.6039908     38          1         123
#> 205   cost:size    effect.power 0.4893051     38          1         123
#> 206   cost:type    effect.power 0.5947261     38          1         123
#> 207   size:type    effect.power 0.8225300     38          1         123
#> 208 (Intercept) parameter.power 0.8131992     38          1         123
#> 209        cost parameter.power 0.7962184     38          1         123
#> 210       size1 parameter.power 0.5104082     38          1         123
#> 211       size2 parameter.power 0.5359919     38          1         123
#> 212       type1 parameter.power 0.3702930     38          1         123
#> 213       type2 parameter.power 0.3654661     38          1         123
#> 214       type3 parameter.power 0.3803947     38          1         123
#> 215  cost:size1 parameter.power 0.4910267     38          1         123
#> 216  cost:size2 parameter.power 0.5103905     38          1         123
#> 217  cost:type1 parameter.power 0.3495862     38          1         123
#> 218  cost:type2 parameter.power 0.3495862     38          1         123
#> 219  cost:type3 parameter.power 0.3773299     38          1         123
#> 220 size1:type1 parameter.power 0.2021133     38          1         123
#> 221 size2:type1 parameter.power 0.2090356     38          1         123
#> 222 size1:type2 parameter.power 0.2032873     38          1         123
#> 223 size2:type2 parameter.power 0.2078892     38          1         123
#> 224 size1:type3 parameter.power 0.2149422     38          1         123
#> 225 size2:type3 parameter.power 0.2152043     38          1         123
#> 226 (Intercept)    effect.power 0.8294224     39          1         123
#> 227        cost    effect.power 0.8143658     39          1         123
#> 228        size    effect.power 0.5315027     39          1         123
#> 229        type    effect.power 0.6352887     39          1         123
#> 230   cost:size    effect.power 0.5131909     39          1         123
#> 231   cost:type    effect.power 0.6153382     39          1         123
#> 232   size:type    effect.power 0.8335248     39          1         123
#> 233 (Intercept) parameter.power 0.8294224     39          1         123
#> 234        cost parameter.power 0.8143658     39          1         123
#> 235       size1 parameter.power 0.5279173     39          1         123
#> 236       size2 parameter.power 0.5441380     39          1         123
#> 237       type1 parameter.power 0.3697391     39          1         123
#> 238       type2 parameter.power 0.3867470     39          1         123
#> 239       type3 parameter.power 0.3932962     39          1         123
#> 240  cost:size1 parameter.power 0.5191018     39          1         123
#> 241  cost:size2 parameter.power 0.5191018     39          1         123
#> 242  cost:type1 parameter.power 0.3548121     39          1         123
#> 243  cost:type2 parameter.power 0.3822490     39          1         123
#> 244  cost:type3 parameter.power 0.3822490     39          1         123
#> 245 size1:type1 parameter.power 0.2044219     39          1         123
#> 246 size2:type1 parameter.power 0.2080000     39          1         123
#> 247 size1:type2 parameter.power 0.2174639     39          1         123
#> 248 size2:type2 parameter.power 0.2198694     39          1         123
#> 249 size1:type3 parameter.power 0.2097721     39          1         123
#> 250 size2:type3 parameter.power 0.2369955     39          1         123
#> 251 (Intercept)    effect.power 0.8368222     40          1         123
#> 252        cost    effect.power 0.8303912     40          1         123
#> 253        size    effect.power 0.5310633     40          1         123
#> 254        type    effect.power 0.6565705     40          1         123
#> 255   cost:size    effect.power 0.5160100     40          1         123
#> 256   cost:type    effect.power 0.6390699     40          1         123
#> 257   size:type    effect.power 0.8417282     40          1         123
#> 258 (Intercept) parameter.power 0.8368222     40          1         123
#> 259        cost parameter.power 0.8303912     40          1         123
#> 260       size1 parameter.power 0.5436906     40          1         123
#> 261       size2 parameter.power 0.5339482     40          1         123
#> 262       type1 parameter.power 0.3985223     40          1         123
#> 263       type2 parameter.power 0.4052082     40          1         123
#> 264       type3 parameter.power 0.3919700     40          1         123
#> 265  cost:size1 parameter.power 0.5269844     40          1         123
#> 266  cost:size2 parameter.power 0.5269844     40          1         123
#> 267  cost:type1 parameter.power 0.3866960     40          1         123
#> 268  cost:type2 parameter.power 0.3900963     40          1         123
#> 269  cost:type3 parameter.power 0.3900963     40          1         123
#> 270 size1:type1 parameter.power 0.2133629     40          1         123
#> 271 size2:type1 parameter.power 0.2365138     40          1         123
#> 272 size1:type2 parameter.power 0.2155276     40          1         123
#> 273 size2:type2 parameter.power 0.2166345     40          1         123
#> 274 size1:type3 parameter.power 0.2172680     40          1         123
#> 275 size2:type3 parameter.power 0.2213500     40          1         123
#> No errors or warnings encountered during power curve generation!